split stb_include into .cpp and .h
